# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "bloggorras Api Documentation",
    "description": "bloggorras Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "bloggorras Logo",
      "altText": "bloggorras Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "bloggorras API Support",
      "url": "https://bloggorras.com/info/support",
      "email": "support@bloggorras.com"
    },
    "termsOfService": "https://bloggorras.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/proveedores": {
      "get": {
        "summary": "List Proveedores",
        "description": "Proveedores",
        "operationId": "ListProveedores",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProveedores"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/proveedores/view/{id}": {
      "get": {
        "summary": "View Proveedores",
        "description": "Ver",
        "operationId": "ViewProveedores",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewProveedores"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/proveedores/add": {
      "post": {
        "summary": "Add Proveedores",
        "description": "Agregar nuevo",
        "operationId": "AddProveedores",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Proveedores"
        ],
        
        "requestBody": {
          "description": "Add Proveedores",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProveedores"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewProveedores"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/proveedores/edit/{id}": {
      "put": {
        "summary": "Edit Proveedores",
        "description": "Editar",
        "operationId": "EditProveedores",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Proveedores",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditProveedores"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditProveedores"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/proveedores/delete/{id}": {
      "delete": {
        "summary": "Delete Proveedores",
        "description": "Proveedores",
        "operationId": "DeleteProveedores",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ListProveedores": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "direccion": {
            "description": "Direccion",
            "type": "string"
          },
          "numero": {
            "description": "Numero",
            "type": "string"
          }
        }
      },
      "ViewProveedores": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "direccion": {
            "description": "Direccion",
            "type": "string"
          },
          "numero": {
            "description": "Numero",
            "type": "string"
          }
        }
      },
      "AddProveedores": {
        "type": "object",
        "properties": {
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "direccion": {
            "description": "Direccion",
            "type": "string"
          },
          "numero": {
            "description": "Numero",
            "type": "string"
          }
        }
      },
      "EditProveedores": {
        "type": "object",
        "properties": {
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "direccion": {
            "description": "Direccion",
            "type": "string"
          },
          "numero": {
            "description": "Numero",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}